:ok_hand: [#824] PR feedback
